requirements.txt
Flask
Flask-Cors
Flask-PyMongo
pymongo[srv]
python-dotenv
dnspython
